:memo: refactoring
