+++ ndlib/models/DynamicCompositeModel.py
